:hammer: Move the BookmarkData class into its own module
